requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
currency_converter
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
gspread-dataframe==3.3.1
idna==3.4
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==1.26.16
